+++ connecteur/test/action/PastellCoreTestFatalError.php
@@ -4,11 +4,8 @@ declare(strict_types=1);
 
 class PastellCoreTestFatalError extends ActionExecutor
 {
-    /**
-     * @throws UnrecoverableException
-     */
     public function go(): void
     {
-        throw new UnrecoverableException('Fatal error');
+        throw new Error('Fatal error');
     }
 }